refactor: CustomRepository관계 구조 변경
